Change post query

+++ app/posts/routes.py
@@ -98,25 +98,26 @@ def playlists():
 @posts.route('/post/<int:post_id>/<action>', methods=['POST'])
 @login_required
 def perform_action(post_id, action):
-    post = Post.query.get_or_404(post_id)
-    if action == 'like':
-        current_user.cast(post, 'like')
-        db.session.commit()
-        return make_response('Success', 200)
-    elif action == 'unlike':
-        current_user.uncast(post, 'like')
-        db.session.commit()
-        return make_response('Success', 200)
-    if action == 'fave':
-        current_user.cast(post, 'fave')
-        db.session.commit()
-        return make_response('Success', 200)
-    elif action == 'unfave':
-        current_user.uncast(post, 'fave')
-        db.session.commit()
-        return make_response('Success', 200)
-    elif action == 'delete' and current_user.is_admin:
-        db.session.delete(post)
-        db.session.commit()
-        flash('The video has been deleted', 'success')
-        return redirect(url_for('main.home'))
+    if (post := Post.query.get(post_id)):
+        if action == 'like':
+            current_user.cast(post, 'like')
+            db.session.commit()
+            return make_response('Success', 200)
+        elif action == 'unlike':
+            current_user.uncast(post, 'like')
+            db.session.commit()
+            return make_response('Success', 200)
+        if action == 'fave':
+            current_user.cast(post, 'fave')
+            db.session.commit()
+            return make_response('Success', 200)
+        elif action == 'unfave':
+            current_user.uncast(post, 'fave')
+            db.session.commit()
+            return make_response('Success', 200)
+        elif action == 'delete' and current_user.is_admin:
+            db.session.delete(post)
+            db.session.commit()
+            flash('The video has been deleted', 'success')
+            return redirect(url_for('main.home'))
+    return make_response('Post not found', 404)
